Move route to corresponse file & create config & optimize import
